modifications emplacements, entre autres
