+++ app/src/main/java/com/marmot/intrepid/naturalhealer/control/MainActivity.java
@@ -1,4 +1,4 @@
-package com.marmot.intrepid.naturalhealer.view;
+package com.marmot.intrepid.naturalhealer.control;
 
 import android.os.Bundle;
 import android.support.v4.app.FragmentTransaction;
@@ -27,6 +27,7 @@ public class MainActivity extends AppCompatActivity
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+
         Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
