+++ app/src/main/java/com/marmot/intrepid/naturalhealer/control/ShopFragment.java
@@ -1,11 +1,14 @@
-package com.marmot.intrepid.naturalhealer.view;
+package com.marmot.intrepid.naturalhealer.control;
 
 import android.content.Context;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
+import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Button;
+import android.widget.ListView;
 
 import com.marmot.intrepid.naturalhealer.R;
 
@@ -13,12 +16,12 @@ import com.marmot.intrepid.naturalhealer.R;
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link InfoFragment.OnFragmentInteractionListener} interface
+ * {@link ShopFragment.OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link InfoFragment#newInstance} factory method to
+ * Use the {@link ShopFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class InfoFragment extends Fragment {
+public class ShopFragment extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -30,7 +33,7 @@ public class InfoFragment extends Fragment {
 
     private OnFragmentInteractionListener mListener;
 
-    public InfoFragment() {
+    public ShopFragment() {
         // Required empty public constructor
     }
 
@@ -40,11 +43,11 @@ public class InfoFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment InfoFragment.
+     * @return A new instance of fragment ShopFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static InfoFragment newInstance(String param1, String param2) {
-        InfoFragment fragment = new InfoFragment();
+    public static ShopFragment newInstance(String param1, String param2) {
+        ShopFragment fragment = new ShopFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -64,17 +67,68 @@ public class InfoFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        View view= inflater.inflate(R.layout.fragment_info, container, false);
-
+        View view = inflater.inflate(R.layout.fragment_shop, container, false);
         // NOTE : We are calling the onFragmentInteraction() declared in the MainActivity
         // ie we are sending "Fragment 1" as title parameter when fragment1 is activated
         if (mListener != null) {
-            mListener.onFragmentInteraction("INFORMATIONS");
+            mListener.onFragmentInteraction("SHOP");
         }
 
         // Here we will can create click listners etc for all the gui elements on the fragment.
         // For eg: Button btn1= (Button) view.findViewById(R.id.frag1_btn1);
         // btn1.setOnclickListener(...
+
+        final ListView list = (ListView) view.findViewById(R.id.listCategories);
+
+        //list.setAdapter(new ArrayAdapter<String>(getActivity(), android.R.layout.simple_list_item_1, categories));
+
+        final Button herbs = (Button) view.findViewById(R.id.buttonHerbs);
+        final Button recipies = (Button) view.findViewById(R.id.buttonRecipies);
+        final Button other = (Button) view.findViewById(R.id.buttonOtherIngredients);
+
+        herbs.setPressed(true);
+
+        herbs.setOnTouchListener(new View.OnTouchListener(){
+            @Override
+            public boolean onTouch(View v, MotionEvent event){
+                if(event.getAction() == MotionEvent.ACTION_DOWN){
+                    recipies.setPressed(false);
+                    other.setPressed(false);
+                    herbs.setPressed(true);
+                    //categories = new String[]{"Aromatiques", "Sauvages", "Légumineuses"};
+                    //list.setAdapter(new ArrayAdapter<String>(getActivity(), android.R.layout.simple_list_item_1, categories));
+                }
+                return true;//Return true, so there will be no onClick-event
+            }
+        });
+
+        recipies.setOnTouchListener(new View.OnTouchListener(){
+            @Override
+            public boolean onTouch(View v, MotionEvent event){
+                if(event.getAction() == MotionEvent.ACTION_DOWN){
+                    herbs.setPressed(false);
+                    other.setPressed(false);
+                    recipies.setPressed(true);
+                    //categories = new String[]{"Tisanes", "Onguents", "Soupes", "Autre chose"};
+                    //list.setAdapter(new ArrayAdapter<String>(getActivity(), android.R.layout.simple_list_item_1, categories));
+                }
+                return true;//Return true, so there will be no onClick-event
+            }
+        });
+
+        other.setOnTouchListener(new View.OnTouchListener(){
+            @Override
+            public boolean onTouch(View v, MotionEvent event){
+                if(event.getAction() == MotionEvent.ACTION_DOWN){
+                    herbs.setPressed(false);
+                    recipies.setPressed(false);
+                    other.setPressed(true);
+                    //categories = new String[]{"Tisanes", "Onguents", "Soupes", "Autre chose"};
+                    //list.setAdapter(new ArrayAdapter<String>(getActivity(), android.R.layout.simple_list_item_1, categories));
+                }
+                return true;//Return true, so there will be no onClick-event
+            }
+        });
 
         return view;
     }
