+++ app/src/main/res/layout/fragment_info.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.marmot.intrepid.naturalhealer.view.InfoFragment">
+    tools:context="com.marmot.intrepid.naturalhealer.control.InfoFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
