+++ app/src/main/res/layout/fragment_inventory.xml
@@ -6,7 +6,7 @@
     android:paddingLeft="10dp"
     android:paddingBottom="10dp"
     android:paddingRight="10dp"
-    tools:context="com.marmot.intrepid.naturalhealer.view.InventoryFragment">
+    tools:context="com.marmot.intrepid.naturalhealer.control.InventoryFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
