+++ app/src/main/res/layout/fragment_map.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.marmot.intrepid.naturalhealer.view.MainFragment">
+    tools:context="com.marmot.intrepid.naturalhealer.control.MainFragment">
 
     <HorizontalScrollView android:id="@+id/hScroll"
         android:layout_width="fill_parent"
