+++ app/src/main/res/layout/fragment_shop.xml
@@ -6,7 +6,7 @@
     android:paddingLeft="10dp"
     android:paddingBottom="10dp"
     android:paddingRight="10dp"
-    tools:context="com.marmot.intrepid.naturalhealer.view.ShopFragment">
+    tools:context="com.marmot.intrepid.naturalhealer.control.ShopFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
